Add model method and test requiring an address to checkout

+++ app/controllers/orders_controller.rb
@@ -8,21 +8,26 @@ class OrdersController < ApplicationController
   end
 
   def create
-    order = current_user.orders.create(order_params)
-    if order.save
-      cart.items.each do |item, quantity|
-        order.item_orders.create(
-          item: item,
-          quantity: quantity,
-          price: item.price,
-          merchant_id: item.merchant_id
-        )
+    if current_user.address?
+      order = current_user.orders.create(address_id: params[:address_id])
+      if order.save
+        cart.items.each do |item, quantity|
+          order.item_orders.create(
+            item: item,
+            quantity: quantity,
+            price: item.price,
+            merchant_id: item.merchant_id
+          )
+        end
+        session.delete(:cart)
+        redirect_to "/orders/#{order.id}"
+      else
+        flash[:notice] = 'Please complete address form to create an order.'
+        render :new
       end
-      session.delete(:cart)
-      redirect_to "/orders/#{order.id}"
     else
-      flash[:notice] = 'Please complete address form to create an order.'
-      render :new
+      flash[:error] = 'An address is required to checkout.'
+      redirect_to '/profile/addresses/new'
     end
   end
